refactor: improve useCombinedRefs code (#27181)
* improve: useCombinedRefs code

* change location
